+++ src/main/java/com/sandymandy/pleasurecraft/screen/client/GirlSceneScreen.java
@@ -1,11 +1,14 @@
 package com.sandymandy.pleasurecraft.screen.client;
 
+import com.sandymandy.pleasurecraft.networking.C2S.InInventoryC2SPacket;
 import com.sandymandy.pleasurecraft.networking.C2S.StartSceneC2SPacket;
+import com.sandymandy.pleasurecraft.screen.InventoryButtonAction;
 import com.sandymandy.pleasurecraft.util.SceneOptions;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.DrawContext;
 import net.minecraft.client.gui.screen.Screen;
+import net.minecraft.client.gui.tooltip.Tooltip;
 import net.minecraft.client.gui.widget.ButtonWidget;
 import net.minecraft.text.Text;
 
@@ -13,11 +16,13 @@ import java.util.List;
 
 public class GirlSceneScreen extends Screen {
     private final int entityId;
+    private final int currentRelationshipLevel;
     private final List<SceneOptions> sceneOptions;
 
-    public GirlSceneScreen(int entityId, List<SceneOptions> sceneOptions) {
+    public GirlSceneScreen(int entityId, int currentRelationshipLevel, List<SceneOptions> sceneOptions) {
         super(Text.literal("Scene Options"));
         this.entityId = entityId;
+        this.currentRelationshipLevel = currentRelationshipLevel;
         this.sceneOptions = sceneOptions;
     }
 
@@ -25,13 +30,23 @@ public class GirlSceneScreen extends Screen {
     protected void init() {
         int y = this.height / 4;
         for (SceneOptions sceneOptions : sceneOptions) {
-            this.addDrawableChild(ButtonWidget.builder(Text.of(sceneOptions.name()), button -> {
+            ButtonWidget buttonWidget = ButtonWidget.builder(Text.of(sceneOptions.name()), button -> {
                 ClientPlayNetworking.send(new StartSceneC2SPacket(
                         this.entityId,
                         sceneOptions
                 ));
                 MinecraftClient.getInstance().setScreen(null); // close after sending
-            }).dimensions(this.width / 2 - 100, y, 200, 20).build());
+            }).dimensions(this.width / 2 - 100, y, 200, 20).build();
+
+            if (this.currentRelationshipLevel < sceneOptions.requiredRelationshipLevel()) {
+                buttonWidget.active = false; // disables and grays out
+            }
+
+            if (!buttonWidget.active) {
+                buttonWidget.setTooltip(Tooltip.of(Text.literal("Requires relationship level " + sceneOptions.requiredRelationshipLevel())));
+            }
+
+            this.addDrawableChild(buttonWidget);
             y += 25;
         }
     }