+++ src/main/java/com/sandymandy/pleasurecraft/util/SceneOptions.java
@@ -1,5 +1,6 @@
 package com.sandymandy.pleasurecraft.util;
 
+import com.sandymandy.pleasurecraft.networking.codec.PacketCodecExtra;
 import net.minecraft.network.RegistryByteBuf;
 import net.minecraft.network.codec.PacketCodec;
 import net.minecraft.network.codec.PacketCodecs;
@@ -15,13 +16,15 @@ public record SceneOptions(
         String cumAnim,
         boolean isBedScene,
         float bedOffset,
-        List<String> bedIdle
+        List<String> bedIdle,
+        boolean needsToStrip,
+        int requiredRelationshipLevel
 
 ) {
 
-    public static final SceneOptions EMPTY = new SceneOptions("",new ArrayList<>(),new ArrayList<>(),new ArrayList<>(),"",false,0f, new ArrayList<>());
+    public static final SceneOptions EMPTY = new SceneOptions("",new ArrayList<>(),new ArrayList<>(),new ArrayList<>(),"",false,0f, new ArrayList<>(), false, 0);
 
-    public static final PacketCodec<RegistryByteBuf, SceneOptions> CODEC = PacketCodec.tuple(
+    public static final PacketCodec<RegistryByteBuf, SceneOptions> CODEC = PacketCodecExtra.tuple(
             PacketCodecs.STRING, SceneOptions::name,
             PacketCodecs.collection(ArrayList::new , PacketCodecs.STRING), SceneOptions::introAnim,
             PacketCodecs.collection(ArrayList::new , PacketCodecs.STRING), SceneOptions::slowAnim,
@@ -30,6 +33,8 @@ public record SceneOptions(
             PacketCodecs.BOOLEAN, SceneOptions::isBedScene,
             PacketCodecs.FLOAT, SceneOptions::bedOffset,
             PacketCodecs.collection(ArrayList::new, PacketCodecs.STRING), SceneOptions::bedIdle,
+            PacketCodecs.BOOLEAN, SceneOptions::needsToStrip,
+            PacketCodecs.VAR_INT, SceneOptions::requiredRelationshipLevel,
             SceneOptions::new
     );
 }